- added latest GGZ Patch by Josef Spillner
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@867 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/militarysite.cc
@@ -139,7 +139,7 @@ std::string MilitarySite::get_statistics_string()
 {
    char str[255];
    if (m_soldier_requests.size())
-      sprintf (str, "%d soldiers (+%d)", m_soldiers.size(), (int)m_soldier_requests.size());
+      sprintf (str, "%d soldiers (+%d)", (int)m_soldiers.size(), (int)m_soldier_requests.size());
    else
       sprintf (str, "%d soldiers", (int)m_soldiers.size());
